requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.2.3
decouple==0.0.7
dj-database-url==2.0.0
Django==4.2.4
django-browser-reload==1.11.0
django-filter==23.2
django-heroku==0.3.1
django-model-utils==4.3.1
django-tailwind==3.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
